fix: fix tooltip
