+++ packages/shared/components/StrongholdWarningBadge.svelte
@@ -1,0 +1,25 @@
+<script lang="ts">
+    import { Text, Tooltip } from './'
+    import WarningBadge from './WarningBadge.svelte'
+    import { localize } from '@core/i18n'
+
+    let strongholdTooltipAnchor = undefined
+    let showStrongholdTooltip = false
+    function toggleTooltip(): void {
+        showStrongholdTooltip = !showStrongholdTooltip
+    }
+</script>
+
+<stronghold-warning-badge
+    bind:this={strongholdTooltipAnchor}
+    on:mouseenter={toggleTooltip}
+    on:mouseleave={toggleTooltip}
+>
+    <WarningBadge />
+</stronghold-warning-badge>
+
+{#if showStrongholdTooltip}
+    <Tooltip anchor={strongholdTooltipAnchor} position="right" size="small">
+        <Text type="p">{localize('views.login.outdatedStronghold')}</Text>
+    </Tooltip>
+{/if}
